API

TriLayout

TriMatrices.TriLayout — Type

abstract type TriLayout

Abstract type which indicates how a TriMatrix's values are layed out in memory. Type parameter is a Bool indicating whether the matrix diagonal is stored.

TriMatrices.TriUpper — Type

struct TriUpper{D} <: TriMatrices.TriLayout{D}

The upper triangle of the matrix is stored, values beneath the diagonal are zero.

TriMatrices.TriLower — Type

struct TriLower{D} <: TriMatrices.TriLayout{D}

The lower triangle of the matrix is stored, values above the diagonal are zero.

TriMatrices.TriSymmetric — Type

struct TriSymmetric{D} <: TriMatrices.TriLayout{D}

Matrix is symmetric across the diagonal, one value is stored for each pair of non-diagonal entries.

TriMatrices.hasdiag — Function

hasdiag(::Type{<:TriLayout})
hasdiag(layout::TriLayout)

Check if the given layout/layout type stores values along the diagonal.

TriMatrices.nelems — Function

nelems(_::Type{L<:TriMatrices.TriLayout}, n::Integer) -> Any

Get the number of elements needed to store the data of a TriMatrix with the given layout.

TriMatrix

TriMatrices.TriMatrix — Type

struct TriMatrix{T, L<:TriMatrices.TriLayout, A} <: AbstractArray{T,2}

A triangular or symmetric matrix which stores data non-redundantly in a contiguous linear array.

Methods

Indexing

TriMatrices.Indexing.check_tri_index — Function

check_tri_index(Bool, layout::TriLayout, i, j)::Bool

Check whether a row/column index is within the stored region of a TriMatrix with the given layout. Assumes the index is otherwise valid for the corresponding matrix size.

TriMatrices.Indexing.car2lin — Function

car2lin(layout::TriLayout, i::Integer, j::Integer)
car2lin(layout::TriLayout, idx::CartesianIndex{2})

Convert Cartesian row/column indices of a TriMatrix with the given layout to a linear index of its data array.

This function is the inverse of lin2car.

TriMatrices.Indexing.car2lin_unchecked — Function

car2lin_unchecked(layout::TriLayout, i::Integer, j::Integer)
car2lin_unchecked(layout::TriLayout, idx::CartesianIndex{2})

Like car2lin, but somewhat faster because it does not check that the index is within the data region of the matrix.

TriMatrices.Indexing.lin2car — Function

lin2car(layout::TriLayout, i::Integer)

Convert linear index of the data array of a TriMatrix with the given layout to the corresponding Cartesian index.

This function is the inverse of car2lin.

TriMatrices.Indexing.tri_indices — Function

tri_indices(layout::TriMatrices.TriLayout, n::Integer) -> TriMatrices.Indexing.TriIndexIterator

Iterate over the cartesian indices of a TriMatrix corresponding to the linear indices of its data array, in order.

This should be equivalent to the following generator expression:

(lin2car(layout, i) for i in 1:nelems(layout, n))

TriMatrices.Indexing.TriIndexIterator — Type

Iterator over Cartesian indices of the stored region of a TriMatrix in an order corresponding to the elements of its data array. Created with the tri_indices function.

TriMatrices.Indexing.trinum — Function

trinum(n::Integer) -> Any

Get the nth triangular number.

See also

triinv

TriMatrices.Indexing.triinv — Function

triinv(t::Integer) -> Any

Get the largest n such that t >= trinum(n).

See also

triinv_strict, trinum

TriMatrices.Indexing.triinv_strict — Function

triinv_strict(t::Integer) -> Any

Get n such that t is the nth triangular number. If t is not a triangular number, throw a DomainError.

See also

triinv, trinum

TriMatrices.Indexing.triinv_rem — Function

triinv_rem(t::Integer) -> Tuple{Any,Any}

Get n and r >= 0 such that t == trinum(n) + r and t < trinum(n + 1).

See also

triinv, trinum