API

TriLayout

TriMatrices.TriLayout — Type

abstract type TriLayout

Abstract type which indicates which regions of a TriMatrix are stored in memory. Type parameter is a Bool indicating whether the matrix diagonal is stored.

TriMatrices.TriUpper — Type

struct TriUpper{D} <: TriMatrices.TriLayout{D}

The upper triangle of the matrix is stored, values beneath the diagonal are zero.

TriMatrices.TriLower — Type

struct TriLower{D} <: TriMatrices.TriLayout{D}

The lower triangle of the matrix is stored, values above the diagonal are zero.

TriMatrices.TriSymmetric — Type

struct TriSymmetric{D} <: TriMatrices.TriLayout{D}

Matrix is symmetric across the diagonal, one value is stored for each pair of non-diagonal entries.

TriMatrices.hasdiag — Function

hasdiag(::Type{<:TriLayout})::Bool
hasdiag(layout::TriLayout)::Bool

Check if the given layout/layout type stores values along the diagonal.

TriMatrices.nelems — Function

nelems(::Type{<:TriLayout}, n)::Integer
nelems(layout::TriLayout, n)::Integer

Get the number of elements needed to store the data of a TriMatrix with the given layout.

TriMatrices.transpose_layout — Function

transpose_layout(::Type{<:TriLayout})
transpose_layout(layout::TriLayout)

Get the TriLayout of the transpose of a TriMatrix with the given layout.

This maps TriUpper and TriLower to each other, and TriSymmetric to itself.

TriMatrix

TriMatrices.TriMatrix — Type

struct TriMatrix{T, L<:TriMatrices.TriLayout, A} <: AbstractArray{T,2}

A triangular or symmetric matrix which stores data non-redundantly in a contiguous linear array.

Missing docstring.

Missing docstring for TriMatrix(::TriMatrices.TriLayout, ::Integer, ::AbstractVector). Check Documenter's build log for details.

TriMatrices.TriMatrix — Method

TriMatrix{T}(layout::TriLayout, undef, n::Integer; diag=zero(T))
TriMatrix(layout::TriLayout, undef, n::Integer; diag=0.)

Create an uninitialized n by n TriMatrix.

T defaults to Float64 if not specified.

TriMatrices.TriMatrix — Method

TriMatrix{T}(layout::TriLayout, m::AbstractMatrix; diag=zero(T))
TriMatrix(layout::TriLayout, m::AbstractMatrix; [diag])

Create a TriMatrix by copying data from a square matrix m.

If unspecified T defaults to eltype(m).

Base.fill — Method

fill(x, layout::TriLayout, n::Integer; diag=x)

Create an n by n TriMatrix with the stored data region filled with the value x.

Base.ones — Method

ones(T::Type=Float64, layout::TriLayout, n::Integer; diag=x)

Create an n by n TriMatrix with the stored data region filled with ones.

Base.zeros — Method

zeros(T::Type=Float64, layout::TriLayout, n::Integer; diag=x)

Create an n by n TriMatrix with the stored data region filled with zeros.

TriMatrices.TriLayout — Method

TriMatrices.TriLayout(m::TriMatrix)

Get the TriLayout of the given TriMatrix instance.

Base.parent — Method

parent(m::TriMatrix)

Get the underlying data array of the TriMatrix m.

TriMatrices.getindex_tri_unsafe — Function

getindex_tri_unsafe(mat::TriMatrix, i::Integer, j::Integer) -> Any

Get the value of an element within the stored data region of a TriMatrix. Should be faster than Base.getindex (even with @inbounds applied) but results are undefined if check_tri_index(mat, i, j) does not pass.

TriMatrices.setindex_tri_unsafe! — Function

setindex_tri_unsafe!(mat::TriMatrix, v::Any, i::Integer, j::Integer) -> Any

Set the value of an element within the stored data region of a TriMatrix. Should be faster than Base.setindex! (even with @inbounds applied) but results are undefined if check_tri_index(mat, i, j) does not pass.

Indexing

TriMatrices.Indexing.check_tri_index — Function

check_tri_index(Bool, layout::TriLayout, i, j)::Bool
check_tri_index(Bool, layout::TriLayout, idx::CartesianIndex{2})::Bool

Check whether a row/column index is within the stored region of a TriMatrix with the given layout. Assumes the index is otherwise valid for the corresponding matrix size.

TriMatrices.Indexing.car2lin — Function

car2lin(layout::TriLayout, i::Integer, j::Integer)
car2lin(layout::TriLayout, idx::CartesianIndex{2})

Convert Cartesian row/column indices of a TriMatrix with the given layout to a linear index of its data array.

This function is the inverse of lin2car.

TriMatrices.Indexing.car2lin_unchecked — Function

car2lin_unchecked(layout::TriLayout, i::Integer, j::Integer)
car2lin_unchecked(layout::TriLayout, idx::CartesianIndex{2})

Like car2lin, but somewhat faster because it does not check that the index is within the data region of the matrix.

TriMatrices.Indexing.lin2car — Function

lin2car(layout::TriLayout, i::Integer)

Convert linear index of the data array of a TriMatrix with the given layout to the corresponding Cartesian index.

This function is the inverse of car2lin.

TriMatrices.Indexing.tri_indices — Function

tri_indices(layout::TriLayout, n::Integer)

Iterate over the cartesian indices of a TriMatrix corresponding to the linear indices of its data array, in order.

This should be equivalent to the following generator expression:

(lin2car(layout, i) for i in 1:nelems(layout, n))

TriMatrices.Indexing.TriIndexIterator — Type

Iterator over Cartesian indices of the stored region of a TriMatrix in an order corresponding to the elements of its data array. Created with the tri_indices function.

TriMatrices.Indexing.trinum — Function

trinum(n::Integer) -> Any

Get the nth triangular number.

See also

triinv

TriMatrices.Indexing.triinv — Function

triinv(t::Integer) -> Any

Get the largest n such that t >= trinum(n).

See also

triinv_strict, trinum

TriMatrices.Indexing.triinv_strict — Function

triinv_strict(t::Integer) -> Any

Get n such that t is the nth triangular number. If t is not a triangular number, throw a DomainError.

See also

triinv, trinum

TriMatrices.Indexing.triinv_rem — Function

triinv_rem(t::Integer) -> Tuple{Any,Any}

Get n and r >= 0 such that t == trinum(n) + r and t < trinum(n + 1).

See also

triinv, trinum